Moved DTO objects to own folder
