+++ backend/App/Authentication/IUserService.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Core.DTO;
 using Core.Entities;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
