+++ backend/App/Controllers/CategoryEntryController.cs
@@ -3,6 +3,7 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using App.Extensions;
 using Common.Events;
+using Core.DTO;
 using Core.Entities;
 using Core.Models;
 using Core.RepositoriesInterfaces;
