+++ backend/Core/DTO/ChangePasswordDTO.cs
@@ -1,4 +1,4 @@
-namespace Core.Entities
+namespace Core.DTO
 {
     public class ChangePasswordDTO
     {
