+++ backend/Core/RepositoriesInterfaces/ICategoryEntryRepository.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Core.DTO;
 using Core.Entities;
 using Core.Models;
