+++ backend/Core/RepositoriesInterfaces/ICategoryRepository.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Core.DTO;
 using Core.Entities;
 using Core.Models;
