+++ backend/Infrastructure/Repositories/CategoryEntryRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Core.DTO;
 using Core.Entities;
 using Core.Models;
 using Core.RepositoriesInterfaces;
