+++ backend/Infrastructure/Repositories/CategoryRepository.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Core.DTO;
 using Core.Entities;
 using Core.Models;
 using Core.RepositoriesInterfaces;
